Began the move of 'peak.interface' to a separate 'protocols' package for ease of distribution independently of PEAK.  Also, removed caching of success/failed adapter lookups, because there was no way for them to be cleared when declarations were made for base classes of a class whose lookup was cached.  (To do caching, I'll have to add a separate cache.)
In the next pass, I'll add the new declaration API that Ty and I designed,
and then change PEAK to use it, factoring out the interim API, and changing
the few remaining introspections in PEAK to work via adapation.  Then, it's
off to documentation, unit tests, and packaging for the 'protocols'
package.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PyProtocols@5 571e12c6-e1fa-0310-aee7-ff1267fa46bd