Adapter factories are now only called with one argument: the object to adapt.  For backward compatibility, any adapter factories that require more than one argument are wrapped in a converter.  It's highly recommended that you transition to one-argument adapters as soon as practical, since using two-argument adapters will cause deprecation warnings in PyProtocols version 1.0 (and causes PendingDeprecationWarnings in 0.9.3).  This change was made for symmetry with Zope and Twisted adapters, as well as Pythonic adapter factories like 'int' et al.
(Note that as a result of this change, 'Adapter' objects no longer have a
'protocol' attribute, and 'StickyAdapter' objects will also lose their
'protocol' attribute in 1.0.)

Also, I restored the previously-removed 'factory' argument to 'adapt()',
but it generates a DeprecationWarning if you use it.  Thus, programs
written for PyProtocols 0.9.2 that use this argument will keep working
until version 1.0, making the 'protocol' attribute the only bit that's
not 100% backward compatible.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PyProtocols@131 571e12c6-e1fa-0310-aee7-ff1267fa46bd